refactor: 엔티티 관련 클래스 model 패키지에 분리
