+++ back-end/src/main/java/com/tdd/backend/post/PostRepository.java
@@ -3,6 +3,8 @@ package com.tdd.backend.post;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
+import com.tdd.backend.post.model.Post;
+
 @Repository
 public interface PostRepository extends CrudRepository<Post, Long> {
 }
