+++ back-end/src/main/java/com/tdd/backend/post/model/AppointmentStatus.java
@@ -1,4 +1,4 @@
-package com.tdd.backend.post;
+package com.tdd.backend.post.model;
 
 public enum AppointmentStatus {
 	PENDING,
