+++ back-end/src/test/java/com/tdd/backend/post/PostControllerTest.java
@@ -19,6 +19,8 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.tdd.backend.option.data.OptionDto;
 import com.tdd.backend.post.data.LocationDto;
 import com.tdd.backend.post.data.SharingDto;
+import com.tdd.backend.post.model.Post;
+import com.tdd.backend.post.model.RideOption;
 import com.tdd.backend.user.User;
 import com.tdd.backend.user.UserRepository;
